dpkg: Cache the result of the cache check process
The anonymous python function in charge of checking whether a package is
in the cache is called several times (due to the multiple times the file is
parsed), resulting in several processes being spawned. The return value of
the first process that returns is put in cache using the bitbake `persist`
API, saving us some time.