meta: Move 'do_fetch' and 'do_unpack' to base-class
Issues:
unpack and fetch can only be used if you inherit dpkg, they should be
available to others i.e. classes and images

Change:
Pull the code out of the dpdk-class and stick it into the base-class. It is the
same code, no changes to it.

Impact:
This patch does not change the behaviour of Isar. It addresses the
issue, preparing for future patches.

Signed-off-by: Henning Schild <henning.schild@siemens.com>