meta/dpkg-debian: Move 'do_install' to new dpkg-base.bbclass
Issue:
The 'do_install' in dpkg can only be used in this class or one derived from
it. But we are about to have another .deb-creating class, prepare for that
and move common code out.

Change:
Move do_install to a generic class that can be inherited by more specific
dpkg-classes.

Impact:
This patch does not change the behaviour of Isar. It addresses the issue
and prepares for another patch.

Signed-off-by: Henning Schild <henning.schild@siemens.com>