meta/dpkg: Move 'do_build' and buildchroot stuff to dpkg-base.bbclass
Issue:
'do_build' contains the steps we need to "enter" and "exit" BUILDCHROOT
to build packages in there. Other classes that build debian packages
might need that as well but can not use it as long as it is in
dkpg.bbclass.

Change:
Pull 'do_build' and all the buildchroot related stuff out of
dpkg.bbclass and into dpkg-base.bbclass. All that remains is the
implementation of the function dpkg_build().

Impact:
This patch does not change the behaviour of Isar. It addresses the issue
and prepares for another patch.

Signed-off-by: Henning Schild <henning.schild@siemens.com>