meta: dpkg-base: rename 'install' to 'deploy_deb'
Issue:
The task name 'install' is a well known name from OE, gentoo, and other
similar build-systems. But in Isar it is used in a completely different
context, that might confuse Isar users. And we can not use the name for
what people would expect, as long as it is taken.

Change:
Rename the task.

Impact:
This patch does not change the behaviour of Isar. It just addresses the
issue.

Signed-off-by: Henning Schild <henning.schild@siemens.com>