meta: dpkg-raw: rename 'populate_package' to 'install'
Issue:
dpkg-raw uses an unfamiliar name for the step where files make it into
${D}, people coming from OE or gentoo could be confused

Change:
Rename the task from to 'install' which in this case fits the
expectations of people coming from other/similar build systems.

Impact:
Everyone that already based recipes on dpdk-raw.bbclass will have to
change the name in these recipes as well.

Signed-off-by: Henning Schild <henning.schild@siemens.com>