isar-image-base & buildchroot: create hooks_multistrap dir with bb flag
Issue:
Bitbake provides a mechanism to create required directories with the
'dirs' flag of the tasks. This feature should be prefered instead of
depending on the shell.

Change:
Use the 'dirs' flag to create the hooks_multistrap directory instead of
using the 'install' shell command in the build task.

Impact:
No negative impact known.

Signed-off-by: Claudius Heine <ch@denx.de>