image: Copy INITRD_IMAGE out with sudo
Issue: do_copy_boot_files can fail to read initrd if certain packages
are installed in the image. One example package causing that issue is
"dropbear", it is probably connected with the initrd being rebuild
during installation. The file will have mode 0600 instead of the
expected 0644.

Fix: use sudo to copy the file

Impact: There is one more sudo in the Isar codebase. People that faced
the issue are now be able to build their images, others will not see a
difference.

Signed-off-by: Henning Schild <henning.schild@siemens.com>