multistrap: Include -updates suite for distros that support it
Add a second default suite to the image and buildchroot multistrap
config. That works for the three debian DISTROs that Isar supports at
the moment. And those use it, raspbian does not use it.

The result is that you should now get the latest dot-release in
buildchroot and the image.

Signed-off-by: Henning Schild <henning.schild@siemens.com>