dpkg-raw: Empty "Depends:" line is not allowed in .deb
Apt-gat fails to install packages if the control-file contains a
"Depends:"-line without any packages listed.

  E: Problem parsing dependency Depends

So append that line to the control-file only if DEBIAN_DEPENDS was set
in the recipe. The user could still specify i.e. " " or anything
invalid, we do not check that. But now our default works.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Reported-by: Claudius Heine <ch@denx.de>