build.sh: Yet another fix for the installation of build deps
The perl code guessing the build-deps of packages can not handle all the
possible syntax that could be found in Build-Depends. And it ignores
Build-Depends-Indep. So there are several issues in it.

This patch uses an official debian-tool "mk-build-deb" to install the build
deps. The tool generates a package which which will Depends: on the
Build-Depends:. Architecture operators, -Indep etc. are handled
correctly.

Signed-off-by: Henning Schild <henning.schild@siemens.com>