Enable proper rebuilds on dependency changes
Install a basichash as signature handler and set the stamp policy to
full - and suddenly we get proper rebuilds of the image and all affected
packages when some recipe down the dependency chain changed!

We are using the legacy bitbake mechanism here as we do not have
setscene machinery like OE. Still good enough for us.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>