Make distro kernel replaceable
Introduce the configuration variable KERNEL_NAME, typically to be
defined by the machine, that selected the kernel that is installed and
can also be depended on / built against.

Create a shadow recipe linux-disto.bb that simply PROVIDES to the
bitbake world distro kernel targets that exist in package form in either
Debian or Raspian.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>