Provide include for easy custom kernel builds
With this include, it becomes almost trivial to replace the distro
kernel with a custom build. You just need to include linux-custom.inc,
specify the source URI, define via S where the source is unpacked to and
provide a defconfig.

To switch to a custom kernel recipe, KERNEL_NAME has to be adjusted in
local.conf or the machine configuration.

Unless explicitly stated via KERNEL_FLAVOR, this tag is derived from
the including recipe name which then needs to follow the scheme
"linux-<flavor>[_version].bb".

The different kernel defconfig file name can be specified via the
KERNEL_DEFCONFIG variable so that also machine-specific defconfigs can
be set in the same recipe (KERNEL_DEFCONFIG_my-machine = "...").

The approach works internally by first running "make deb-pkg" on the
kernel and the repackages the output to make the binary linux-image and
linux-header debs act as proper replacement of a distro kernel. This
results in a suboptimal technical implementation which may eventually be
replaced by an isar-implemented deb-pkg build process. However, this is
not expected to affect the user-visible interface of this class.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>