Make sure git repos are usable inside buildchroot
bitbake clones git repos into the package workdir in shared mode. This
creates an alternates link to the original download folder. However,
that path is invalid inside buildchroot. This disturbs package builds
which try to derive information from the git repo, such as tags or the
delta to the last release.

Bind-mount the downloads folder and patch the alternates file to restore
the git repo access inside buildchroot.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>