Avoid name collisions of deployed kernels and initrds
While distro kernels tend to have individualizing version names, this
need not be the case for custom kernels. If we deploy them as is into
the common image folder, they can easily overwrite each other when
building images for different distro-machine tuples.

Avoid this by appending <distro>-<machine> to the deployed image names.
Take this into account when using the images in the start_vm script.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>