meta-isar/isar-image-base: Remove uncesessary cmd lines from cfg scripts
Since we changed to use debootstrap and apt for the installation of
every package, additional runs of 'dpkg configure' and postinst scripts
are no longer necessary.

This removes also the defunct revert of the 'daemon disable steps' used
with the multistrap version of isar and the setting of the root
password, since that should be done in a postinst step of a package
instead.

Signed-off-by: Claudius Heine <ch@denx.de>