meta-isar/isar-image-base: Make kernel optional
Isar should also be able to build root file systems without a kernel,
for instance when using it via nfs or on systems where the kernel needs
to be deployed some where else, not as part of the directory tree.

With this patch setting "KERNEL_NAME" to an empty string or not setting
this variable at all, will not cause an error and no kernel will be
installed to the root file system.

Signed-off-by: Claudius Heine <ch@denx.de>