isar-bootstrap: Remove leaked hostname and resolv.conf from images
debootstrap will leak the build hosts /etc/hostname and /etc/resolv.conf
into all rootfss it builds. That is done so the newly created rootfs
will have internet access once you chroot into it.

For the buildchroot we need internet and the leakage does not hurt, for
the final image we probably do not want any of these files anymore.

So split up the apt-get into a download and install phase and delete
the two files after fetching the packages, but only for the image and
not the buildchroot.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Reviewed-by: Claudius Heine <ch@denx.de>