Correctly use the bitbake variable S from now on
S is defined as ${WORKDIR}/${P} so it already contains ${WORKDIR} by its
definition. In Isar it was often concatinated with WORKDIR again. One
example where this was a problem is if you specified a patch in SRC_URI
but did not actually overwrite S.
Align the use of the variable with OE and bitbake defaults again.

Impact: layers building on top of Isar will have to adjust their recipes
like the internal ones needed modification. The suggestion is to not set
S and make sure to unpack to ${WORKDIR}/${P}. When setting S make sure
to include ${WORKDIR}. This patch also introduces a warning and tries to
preserve the old behaviour a bit. However if your recipe uses patches in
SRC_URI and sets S you will get the warning and do_patch will fail.

Signed-off-by: Henning Schild <henning.schild@siemens.com>