scripts/vm_smoke_test: Add quite parameter
Due to this script is used for QA needs only, it's OK that it has verbose
output enabled by default. This patch adds option to supress excessive
output. But the boot log for specific configuration will be anyway shown
if respective test case failed.

Also this patch fixes an issue that the same boot log could be displayed
multiple times.

Signed-off-by: Alexander Smirnov <asmirnov@ilbers.de>