doc: Update sudo configuration in the manual
Francesco reported build problems with the suggested sudo configuration.
The issue might be that somewhere along the way a user or group other
than root is used, or that there are problems with (NO)SETENV.

Update the manual to include an entry that should prevent such problems,
also add a line to deal with proxies.

Reported-by: Francesco <fdprnd@dave.eu>
Signed-off-by: Henning Schild <henning.schild@siemens.com>