meta-isar: wic: Temporarily reset forked wic files to where we forked
Issue: When looking at what we changed in the plugins we could have a
hard time spotting Isar-specific changes, especially when updating wic
some day.

This patch resets the forked files to be exact copies of our current wic
version. This commit breaks wic support in Isar and should only be used
with the following revert. Its only purpose is to hopefully make life
easier in the future.

Signed-off-by: Henning Schild <henning.schild@siemens.com>