isar-bootsrtap: Setup locale
Currently locale is set in config scripts for buildchroot and images.
This is absolutely the same routine which is duplicated in all these
scripts. This patch moves locale initialization to isar-bootstrap, so
now locale is set once and works for all the futher root filesystems.

The task do_set_locale is added right after do_bootstrap to avoid warning
messages in isar-bootstrap logs after 'apt-get' calls.

Signed-off-by: Alexander Smirnov <asmirnov@ilbers.de>