meta-isar/isar-image-base: Added removal of qemu-*-static binaries
After the image has finished building, the qemu-*-static binaries that
are installed by quemu-debootstrap can be removed because they are no
longer needed.

If this binary is overwritten by installing qemu-user-static into the
image, then this binary is not removed. (This is possible if the
binfmt_misc entry for this architecture has the 'F' flag. In that case
qemu-*-static shouldn't even be deployed by qemu-debootstrap.)

Reported-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Claudius Heine <ch@denx.de>