meta/isar-cfg-localepurge: added locale purge recipe
The package 'isar-cfg-localepurge' is used to set the locales of the
root file system and purge all not required locales.

The variables used to configure the locale with this package are `LOCALE_GEN`
which is basically the content of the desired `/etc/locale.gen` with each
line seperated by '\n' and `LOCALE_DEFAULT` which sets the global default
locale of the system.

Signed-off-by: Claudius Heine <ch@denx.de>