buildchroot: Break up wic build deps into common and arch-specific ones
wic is currently not usable on non-x86 archs because of naturally
unresolvable dependencies on x86 grub-efi packages. Fix this by
factoring our a common dependency core and a set of arch-specific ones.
This should also allow to build EFI images for armhf and arm64
(Tianocore firmware or U-boot in EFI mode). But primarily, it allows to
build non-EFI images on non-x86 archs via wic.

Note that we should probably consider replacing the Isar-proprietary
BUILDCHROOT_PREINSTALL_WIC variable with the OE standard
WIC_FILE_DEPENDS*.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>