meta/image: leave BUILD_ID and VARIANT in /etc/os-release
This commit adds support for leaving a mark in the resulting image.
That should allow you to tell what code you used to create it. By
default "git describe" will be used on Isar itself to come up with a
BUILD_ID. And the DESCRIPTION of the image will become the VARIANT.

see
https://www.freedesktop.org/software/systemd/man/os-release.html

Signed-off-by: Henning Schild <henning.schild@siemens.com>