meta/image: allow customization of command we use fo BUILD_ID
One of the previous patches introduced the feature to leave some build
system information in the rootfs. The git command to derive that
information was kept pretty repository agnostic.

This patch makes the command Isar-specific and allows users to customize
the command to their needs if they use a different tagging scheme in
their repository, or if they do not use git at all.

Signed-off-by: Henning Schild <henning.schild@siemens.com>