buildhcroot/build.sh: Disable deps checking
This change is needed in order to avoid failures of dpkg-buildpackage
when a package provides python-pip:native or python-mako:native as
build-depends. ":native" would make the builddeps check fail while that
is needed to install (via mk-build-deps) the host-side version of
these packages in a cross-build case.

Initially it was proposed to use '--no-check-builddeps' parameter, but
seems dpkg-buildpackage from Jessie doesn't support it. I've found,
that '-d' means the same and it's supported in all the suites.

Signed-off-by: Alexander Smirnov <asmirnov@ilbers.de>
Acked-by: Jan Kiszka <jan.kiszka@siemens.com>