doc: clarify usage of the IMAGE_TYPE variable
Documentation of the IMAGE_TYPE variable was misleading since it
claimed that the variable could hold a space separated list of
image types to be generated. The current image generation code
really assumes that a single type will be specified. Also update
the documentation to list "wic-img" as a provided IMAGE_TYPE.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>