linux: add "git" to KBUILD_DEPENDS
Several build scripts from the kernel use "git" and in particular
scripts/setlocalversion. The latter is used to produce a version
string when we "make -s kernelrelease". My kernel builds would
sporadically fail with the PV check as the "+" sign appended by
kbuild would or would not be present in the "kernelrelease" output
if "git" was installed or not in the buildchroot. Get consistent
results by getting "git" installed into the buildchroot when we
build custom kernels.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>