linux-module: Run depmod for installed kernel versions
The return code of the postinst script it the one of the last executed
instruction. Unfortunately, depmod tend to fail when run on the host
during rootfs generation because host kernel and target kernel usually
differ.

To overcome this, pick up the kernel version(s) by looking at the
content of /lib/modules.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>