Install wic and bootloader dependencies via imager_install_deps
This removes the wic dependency installation hack from
buildchroot-target, modeling this instead with the new IMAGER_INSTALL
interface. The basic dependencies of wic are still automatically
installed as soon as the image type is switched to wic-img. However,
those of the bootloader now requires explicit statements.

For that, GRUB_BOOTLOADER_INSTALL and SYSLINUX_BOOTLOADER_INSTALL
variables are introduced and set by the distro confs accordingly. So the
user just needs to assign their value to IMAGE_INSTALL if the
corresponding bootloader is selected in a WKS file - or by some other
imaging mechanism.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>