meta: Add custom U-Boot recipe include
This simplifies the building of custom U-Boot versions significantly.
Just include u-boot-custom.inc in a recipe, define the source as well as
U_BOOT_CONFIG and the target U_BOOT_BIN (binary). The latter will be
packaged and can then be deployed via IMAGER_INSTALL + IMAGER_BUILD_DEPS
to the buildchroot for installation by the imager (e.g. wic).

The package structure this generates is similar though no identical to
Debian upstream:
 - u-boot-${MACHINE}: contains U_BOOT_BIN
 - u-boot-${MACHINE}-dev: contains tools/env/lib.a as libubootenv.a
   which is needed by SWUpdate (recipe for the latter will eventually be
   contributed)
 - u-boot-tools which replaces the upstream package (minus some special
   SOC-specific tools we skip for simplicity reasons)

There is no plain "u-boot" package as we have no use case for it here.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>