u-boot-script: Become independent of target /etc/fstab
It's a bit hacky to rely on that file during image creation,
specifically as imagers such as wic may change them. So far we used the
side channel ROOTFS_DEV which needs to be maintained separately from the
commonly used source: the wks file of the target.

This changes the interface of /etc/default/u-boot-script slightly and
reworks how its recipe fills that config file:
 - ROOT_PARTITION variable is added, saving the root partition number
 - KERNEL_ARGS replaces KERNEL_ARGS_APPEND and is supposed to include
   root=<rootdev> from now on
 - do_install derives both the root partition number as well as the root
   device from the wks file, when available
 - in all other (special) cases, u-boot-script should be set by a
   customization package

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>