isar-image-base: Disable rootfs check by default
While the fsck hook of initramfs-tools does evaluate our generated fstab
it ignores the filesystem type we provide and rather tries to probe
that. Naturally, that is unreliable in the build environment.

To make things simpler and better configurable, disable the rootfs
check. We will add a helper recipe to reenable this and more, but that
only during first boot when we have the final fstab and and the real
rootfs device available.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>