linux-custom: Simplify and clarify KERNEL_NAME-to-recipe matching
The old mechanism tried to be too smart my deriving missing elements
when defining the PROVIDES of a custom kernel recipe from the currently
selected target KERNEL_NAME. But that broke when there was both a
linux-foo and a linux-foo-bar.

Let's make it straight and simple: A custom kernel recipe can control
the suffix "[-featureset]-flavor" of the generated linux-{image,headers}
meta packages by

 a) calling itself "linux-featureset-flavor"
 b) setting the variable KERNEL_NAME_PRODVIDED to "featureset-flavor"

Effectively, KERNEL_NAME is now matched against KERNEL_NAME_PRODVIDED
(in practice, the meta package names are matched). Actually, this is the
same logic that was already used for distro kernels.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>