meta: Add enable-fsck helper recipe
This is a run-once systemd service that enables checks for all regular
filesystems on the target. We do this on the live system rather during
build because we only have the full information required for
populating the initramfs with the right tools available then - at least
as long as we do not want to patch much more.

Enable this feature in the demo local.conf by default because we have
read/write filesystems here.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>