linux-module/autoload: allow defining multiple modules with custom name
Currently `AUTOLOAD` only supports binary operation that defines if a
module (with the same name as the recipe) should be loaded or not. This
is very cumbersome to work with if one recipe allows deployment of
multiple modules or if the module name differs from the recipe name.

With this patch `AUTOLOAD` will have a new meaning, it revers to a
bitbake array that contains the module names.

This commit breaks the usage of `AUTOLOAD`. Recipes will need to replace
all their:

    AUTOLOAD = "1"

lines with:

    AUTOLOAD = "${PN}"

And their:

    AUTOLOAD = "0"

lines with:

    AUTOLOAD = ""

in order to migrate to the new interface.

Signed-off-by: Claudius Heine <ch@denx.de>