wic-img: Allow chown to fail
f13b2bf71dd841 introduced chowning of possible python cache files. This
operation can fail if the file system where Isar resides is read-only,
in which case there will also be no python cache files.

Simply allow this step to fail, even other failures would not be
problematic because it is a convenience thing to execute the chown.

Signed-off-by: Henning Schild <henning.schild@siemens.com>