meta: Add bbclass to retrieve selected WKS file
This implements a lookup service for the target's WKS_FILE. The goal is
to concentrate the search in a single place in order to provide a
consistent logic.

In addition to permitting absolute paths, get_wks_full_path() implements
almost all search paths of OE:

 - <BBPATH>/wic
 - <BBPATH>/scripts/lib/wic/canned-wks
 - <COREBASE>/scripts/lib/wic/canned-wks

It does not support THISDIR for wic-image recipes because that would
restrict the visibility of that file to a single recipe while we would
like to use it freely, also in other recipes such as u-boot-script.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>