image: include image name in the kernel/initrd image copies
The kernel and initrd images are really image-specific (especially the later
as the initrd is created/updated as packages get installed into the root
file-system). Make sure we retain a per-image copy of these images in the
image deploy directory by inserting the name of the image (${PN}) into the
full name of the artifacts (set in ${KERNEL_IMAGE} and ${INITRD_IMAGE}).

In addition, an IMAGE_FULLNAME variable is introduced to combine the three
components of the image name: the (base) name of the image, the name of the
distribution and the name of the machine. This variable is now used in the
start_vm script and the image classes (image, ext4-img and wic-img).

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>