buildchroot: build debian packages as "builder" not "root"
We used to build packages as "root" and now do that as a regular user.
Not building as "root" allows us to find mistakes in debian/rules where
privileged operations are used while they should not (a sudo was
found in a rules-file). Further some build steps might actually expect
to not run as root (seen in openssl test suite).

Not building as root should increase overall quality and brings us
closer to how debian packages are build by others.

Signed-off-by: Henning Schild <henning.schild@siemens.com>