base-apt: Fix distro and codename retrieval
So far, the base-apt-helper derived the base distro name and codename
(so far called suite, but it was already the codename) from the DISTRO
variable. That broke downstream layers which defined their own DISTRO.

A better approach is to let the distro config file define those
parameters so that they can reliably propagate, also via derived
distros, to the helper tasks.

Fixes: d702b4cceb64 ("base-apt: Add helper class")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>