meta/dpkg-raw: build raw packages like all others
Instead of just packaging a bunch of files, debian/-ize them and send
them through the whole build chain. That was inspired by the chown and
"sudo rm" changes that we recently had to make.
This way is much cleaner than what we did before, and we get all debian
quality assurance on top. But the original idea is to bring random files
into the rootfs, where we might not be able to fulfill the quality. So
there is a way for a recipe to bring its own rules and disable
debhelpers it does not like.

The code is intentionally split into many functions, because in fact
you can use that to debian/-ize random sources as well. But i do not
suggest to try that with this version.

Signed-off-by: Henning Schild <henning.schild@siemens.com>