wic: Add rootfs-u-boot
This implements the translation of the target wks into an U-Boot script
file as wic plugin. It allows to remove any image or kernel specifics
from the u-boot-script package. And then, in turn, permits to make wks
files image-specific again, just like in OE.

As u-boot-script is already installed on the target, we chroot into that
rootfs in order to run update-u-boot-script in-place. In cross-arch
scenarios, we just need to restore qemu-*-static temporarily for this
because do_rootfs already cleaned it up.

The plugin allows to inject the SCRIPT_PREPEND and NO_INITRD parameters
via --sourceparams, which also obsoletes the need to carry /etc/default/
u-boot-script in customization packages.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>