sshd-regen-keys: Fix sshd deadlock on boot
Currently, when sshd-regen-keys runs dpkg-reconfigure, this
will lead to a call to `systemctl restart ssh`.  This call blocks
forever because of course the sshd-regen-keys unit, which is a
dependency of sshd, hasn't finished at this point and can't do so
because it is waiting as well.

To circumvent this deadlock, this commit changes sshd-regen-keys'
behavior so sshd is first disabled and only reenabled after the
job is done.

Signed-off-by: Harald Seiler <hws@denx.de>