buildchroot: Perform upgrade after build dependency installation
When we partially rebuild after updating the version of a build
dependency of another package, we do not properly upgrade that
dependency in buildchroot. The reason is that "apt-get install"
performs upgrades only for the explicitly listed packages. But we
install build dependencies indirectly, we a meta package's dependency.

Resolve that be running an explicit "apt-get upgrade" after the
build dependency installation. This will ensure pulling the latest
versions for isar-apt.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>