linux-custom: skip linux-libc-dev deployment on downgrade
A custom kernel usually deploys linux-libc-dev in the kernels version.
But some packages in the distribution depend on >= expected upstream
version of that package. If we deploy a linux-libc-dev in a lower
version, installing packages like build-essential will not be possible
anymore.
That is especially problematic since incremental builds will pull that
into our buildchroot and break package installation in there.

If detecting a downgrade, do not deploy the package and issue a warning.

Signed-off-by: Henning Schild <henning.schild@siemens.com>