meta: add isar-cfg-rootpw recipe for setting root password
The isar-cfg-rootpw recipe is a central point to set the root password
for images. It provides the `CFG_ROOT_PW` and `CFG_ROOT_LOCKED`
variables, that can be set from any `.conf` file or via
`isar-cfg-rootpw.bbappend`.

This package is installed as a transient package to avoid leaking
passwords set by it via the scripts in `/var/lib/dpkg/info/`.

The `CFG_ROOT_PW` variable contains the root password as crypt(3)
encrypted string. If this variable is empty, root login
is passwordless.

The `CFG_ROOT_LOCKED` variable can be set to "1" in order to lock
the root account, other values leave the account unlocked. Unlocking the
account at a later point will restore the password set by `CFG_ROOT_PW`.

Signed-off-by: Claudius Heine <ch@denx.de>