wic-img: import support for .wks.in files from oe-core
OpenEmbedded supports use of variables in wks files if they are
suffixed with .wks.in (instead of .wks). Use of this suffix adds
a task to the build process to expand variables to their values
when known or to the empty string when they are not. The result
of the evaluation is placed in a regular .wks file placed in the
build folder.

The .wks file for the de0-nano-soc machine was modified to show
use of this feature and get it regularly tested as part of our
regular Isar builds.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>