dpkg-base: derive "Package" and "Architecture" from .deb in cleanup
Instead of using "sed" pattern magic ask the .debs their name and
architecture and remove exactly that. Especially the hardcoded
Architecture caused problems for packages of architecture "all". If
your recipe created such a package it could never be cleaned because it
was added to all the archs of the repo, and later cleaned only from one.
Now if that package changes because you are still working on the recipe,
adding the package again would cause checksum mismatches for all but the
current architecture.
Remove the package from all archs when arch is "all".

Reported-by: Vladislav Jasek <vladislav.jasek@siemens.com>
Signed-off-by: Henning Schild <henning.schild@siemens.com>