doc: make absolute chapter links relative
All the chapter links pointed to "master" and not the documentation of
the commit you actually looked at. So when browsing next you would go
back to old version of the manual when clicking the links. Maybe back to
a time where a whole chapter did not even exist.
Fix that by making those links relative to the current version, whatever
that may be.

Signed-off-by: Henning Schild <henning.schild@siemens.com>