buildchroot: do not assume group of "builder" to be "builder"
We try to align the user and group ids of the user running bitbake with
the users inside the chroots. If that outside user happens to be in a
group that already exists in a fresh installation (i.e. users:100) we
will not create a group called "builder" so checking the group name
later will fail. Just check that the gid is the gid of "builder" instead
of checking that the group name is "builder".

Reported-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Henning Schild <henning.schild@siemens.com>