Remove all uses of subprocess.call(shell=True)
Using shell-expansion in subprocess calls has a lot of nasty
side-effects that are hard to debug should one ever surface.
This commit changes all uses of subprocess to

A) not use shell=True by giving a list of args instead of
   a command-line.
B) ensure return-codes are handled appropriately.

Furthermore, in some places the subprocess usage was changed to
be more idiomatic.

Co-authored-by: Claudius Heine <ch@denx.de>
Signed-off-by: Harald Seiler <hws@denx.de>