Use modern python formatting
The use of % for formatting is discouraged because it has side-
effects that are not immediately obvious.  This commit refactors
all uses of % to a better formatting style.

As f-Strings are not availible in in some python versions we support
(<3.6) the formatter of choice is .format() or manual concatenation
in cases where it is more concise.

This commit additionally refactors the showdata and listtasks tasks
to make better use of the new formatter code.

Signed-off-by: Harald Seiler <hws@denx.de>