debianize: use date of the most recent file instead of "now"
The changelog generated by Isar would use the current date/time
causing each build to be different. Use the date/time of the
most recent file to achieve reproducible builds (assuming files
contained in the payload have the same contents and timestamps).
In the event where the package is not shipping any files, Isar
will default to the current date/time.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>