Make patch task repeatedly applicable
This allows to rerun the patch task even if it was applied already. This
is achieved by recording the executed patch commands and storing them,
along with the original patches, in ${S}/.applied_patches and applying
them in reverse order before applying new patches. If unpack ran prior
to that, .applied_patches was deleted, and nothing is incorrectly
reverted.

Alternatives to this self-written solution would have been using quilt,
which would have added almost 30 MB additional dependencies to our build
environment, or even using the OE implementation for the patch task,
which would have required importing and adjusting a significant amount
of code with only little gain.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>